Homework-7, v03
