+++ main.js
@@ -1,5 +1,5 @@
 import Pokemon from './Pokemon.js';
-import {random, $querySel, $createElem} from './utils.js';
+import { random, $querySel, renderLog } from './utils.js';
 
 const pikachu = new Pokemon({
   name: 'Pikachu',
@@ -17,7 +17,7 @@ const charmander = new Pokemon({
 
 const $btn = $querySel('#btn-kick');
 const $btnKickEnemy = $querySel('#btn-kick-enemy');
-const $logs = $querySel('#logs');
+
 
 function startGame() {
   console.log('Start Game!');
@@ -52,33 +52,6 @@ function startGame() {
       el.textContent = `${textContent} (${counter})`;
       return counter;
     }
-  }
-  
-  function generateLog(pikachu, charmander, count) {
-    let { name, hp: { current, total } } = pikachu;
-    let { name: nameEnemy } = charmander;
-  
-    const logs = [
-      `${name} поперхнулся, и за это ${nameEnemy} с испугу приложил прямой удар коленом в лоб врага. -${count}, \[${current}\/${total}\]`,
-      `${name} вспомнил что-то важное, но неожиданно ${nameEnemy}, не помня себя от испуга, ударил в предплечье врага. -${count}, \[${current}\/${total}\]`,
-      `${name} забылся, но в это время наглый ${nameEnemy}, приняв волевое решение, неслышно подойдя сзади, ударил. -${count}, \[${current}\/${total}\]`,
-      `${name} пришел в себя, но неожиданно ${nameEnemy} случайно нанес мощнейший удар. -${count}, \[${current}\/${total}\]`,
-      `${name} поперхнулся, но в это время ${nameEnemy} нехотя раздробил кулаком \<вырезано цензурой\> противника. -${count}, \[${current}\/${total}\]`,
-      `${name} удивился, а ${nameEnemy} пошатнувшись влепил подлый удар. -${count}, \[${current}\/${total}\]`,
-      `${name} высморкался, но неожиданно ${nameEnemy} провел дробящий удар. -${count}, \[${current}\/${total}\]`,
-      `${name} пошатнулся, и внезапно наглый ${nameEnemy} беспричинно ударил в ногу противника -${count}, \[${current}\/${total}\]`,
-      `${name} расстроился, как вдруг, неожиданно ${nameEnemy} случайно влепил стопой в живот соперника. -${count}, \[${current}\/${total}\]`,
-      `${name} пытался что-то сказать, но вдруг, неожиданно ${nameEnemy} со скуки, разбил бровь сопернику. -${count}, \[${current}\/${total}\]`
-    ];
-  
-    return logs[random(logs.length - 1)];
-  }
-  
-  function renderLog(pikachu, charmander, count) {
-    const $li = $createElem('li');
-    $li.classList.add('logs__item');
-    $li.textContent = charmander ? generateLog(charmander, pikachu, count) : generateLog(pikachu, charmander, count);
-    $logs.insertBefore($li, $logs.children[0]);
   }
 }
